refactor: code refactor between accounts and transactions
