+++ src/repositories/accounts.js
@@ -1,6 +1,16 @@
 import { prisma } from '../lib/prisma.js';
 
 export class AccountsRepository {
+  static async getBalance(accountID) {
+    const account = await prisma.bank_Accounts.findUnique({
+      where: {
+        id: accountID,
+      },
+    });
+
+    return account.balance;
+  }
+
   static async getAccounts(pagination) {
     const accounts = await prisma.bank_Accounts.findMany({
       skip: pagination.offset,
@@ -97,25 +107,6 @@ export class AccountsRepository {
     return account;
   }
 
-  static async accountTransaction(accountID) {
-    const accountTrx = await prisma.transactions.findMany({
-      where: {
-        OR: [
-          {
-            source_account_id: accountID,
-          },
-          {
-            destination_account_id: accountID,
-          },
-        ],
-      },
-    });
-
-    const haveTransaction = accountTrx.length === 0 ? false : true;
-
-    return haveTransaction;
-  }
-
   static async deleteAccount(accountID) {
     const account = await prisma.bank_Accounts.delete({
       where: {
@@ -124,5 +115,32 @@ export class AccountsRepository {
     });
 
     return account;
+  }
+
+  static async updateBalance(accountID, newBalance) {
+    const account = await prisma.bank_Accounts.update({
+      where: {
+        id: accountID,
+      },
+      data: {
+        balance: newBalance,
+      },
+      include: {
+        Users: {
+          include: {
+            Profiles: true,
+          },
+        },
+      },
+    });
+
+    return account;
+  }
+
+  static async insufficient(accountID, amount) {
+    const balance = await this.getBalance(accountID);
+    const insufficient = amount > balance ? true : false;
+
+    return insufficient;
   }
 }
