+++ src/routes/transactions.js
@@ -1,11 +1,7 @@
 import express from 'express';
 import Validator from '../utils/validator.js';
 import { TransactionsController } from '../controllers/transactions.js';
-import {
-  depositSchema,
-  transferSchema,
-  withdrawalSchema,
-} from '../utils/validationSchema.js';
+import { transferSchema } from '../utils/validationSchema.js';
 
 const router = express.Router();
 
@@ -14,11 +10,4 @@ router.route('/:transactionID').get(TransactionsController.getTransaction);
 router
   .route('/')
   .post(Validator(transferSchema), TransactionsController.transfer);
-router
-  .route('/withdrawal/:accountID')
-  .put(Validator(withdrawalSchema), TransactionsController.withdrawal);
-router
-  .route('/deposit/:accountID')
-  .put(Validator(depositSchema), TransactionsController.deposit);
-
 export default router;
