+++ src/services/accounts.js
@@ -1,6 +1,8 @@
 import { ErrorHandler } from '../middlewares/error.js';
 import { AccountsRepository } from '../repositories/accounts.js';
+import { TransactionsRepository } from '../repositories/transactions.js';
 import { UsersRepository } from '../repositories/users.js';
+import { formatRupiah } from '../utils/formatRupiah.js';
 
 export class AccountsService {
   static async register(data) {
@@ -77,7 +79,7 @@ export class AccountsService {
     }
 
     const haveTransaction =
-      await AccountsRepository.accountTransaction(accountID);
+      await TransactionsRepository.accountTransaction(accountID);
 
     if (haveTransaction) {
       throw new ErrorHandler(
@@ -89,5 +91,68 @@ export class AccountsService {
     const deletedUser = await AccountsRepository.deleteAccount(accountID);
 
     return deletedUser;
+  }
+
+  static async deposit(accountID, amount) {
+    const account = await AccountsRepository.getAccountById(accountID);
+
+    if (!account) {
+      throw new ErrorHandler(404, `account with ID: ${accountID} is not found`);
+    }
+
+    const newAccountBalance = parseFloat(account.balance) + amount;
+
+    const deposit = await AccountsRepository.updateBalance(
+      accountID,
+      newAccountBalance,
+    );
+
+    delete deposit.Users.password;
+
+    // add trx data to tbl_trx
+    await TransactionsRepository.addToTransaction(accountID, accountID, amount);
+
+    const trx = {
+      amount: await formatRupiah(amount),
+      currentAccountBalance: deposit,
+    };
+
+    return trx;
+  }
+
+  static async withdrawal(accountID, amount) {
+    const account = await AccountsRepository.getAccountById(accountID);
+
+    if (!account) {
+      throw new ErrorHandler(404, `account with ID: ${accountID} is not found`);
+    }
+
+    const insufficient = await AccountsRepository.insufficient(
+      accountID,
+      amount,
+    );
+
+    if (insufficient) {
+      throw new ErrorHandler(400, `account remaining balance is insufficient`);
+    }
+
+    const newAccountBalance = parseFloat(account.balance) - amount;
+
+    const withdrawal = await AccountsRepository.updateBalance(
+      accountID,
+      newAccountBalance,
+    );
+
+    delete withdrawal.Users.password;
+
+    // add trx data to tbl_trx
+    await TransactionsRepository.addToTransaction(accountID, accountID, amount);
+
+    const trx = {
+      amount: await formatRupiah(amount),
+      currentAccountBalance: withdrawal,
+    };
+
+    return trx;
   }
 }
